Query plan output has been cleaned up and now looks like normal JSON with no formatting and typical schema self-describing shape
The alerts section tells you about table scans or materializations that are not supposed to happen and the plans are in their own section.
Particular types of noteworthy operations go in the stats section.  These have flexible schema so it's just an array of tuples.

```json
{
"alerts" : [

],
"plans" : [
  {
   "query" : "INSERT INTO my_table(str) VALUES(\"Hello from CQL.\"), (\"Edit as you please.\")",
   "stats" : [
      ["SCAN", 1]
    ],
   "plan" : "QUERY PLAN\n|..SCAN 2 CONSTANT ROWS"
  },
  {
   "query" : "SELECT *\n  FROM my_table",
   "stats" : [
      ["SCAN", 1]
    ],
   "plan" : "QUERY PLAN\n|..SCAN TABLE my_table"
  }
]
}
```

The playground docs have also been updated to show this example jq command:

```bash
 ./go_query_plan.sh | jq '.plans[] | {query:.query, plan:.plan}'
```

Which yields

```json
{
  "query": "INSERT INTO my_table(str) VALUES(\"Hello from CQL.\"), (\"Edit as you please.\")",
  "plan": "QUERY PLAN\n|..SCAN 2 CONSTANT ROWS"
}
{
  "query": "SELECT *\n  FROM my_table",
  "plan": "QUERY PLAN\n|..SCAN TABLE my_table"
}

This is technically a breaking change but the old format was kind of weird.

A bit of work was done to make sure there are not extra empty records. There were many such things previously so that the code didn't have to think about
if a trailing comma was needed or not.