When creating the query plan output, if we find any native function calls (via proc or otherwise) we have to replace them with constants.
Since it's QP output the only way this can happen is if the items are in the arg list of a shared fragment, those are evaluated in native context.

We have to make sure that we don't replace normal UDFs or select function and also we don't replace any shared_fragment expression, those are inlined.

Otherwise we use our existing logic to emit the right kind of '1', which was hoisted out for just his purpose.