Resolved arguments/locals should copy sem_type by value not reference
When we resolve argument and local bundles, we need to copy the sem_type of the original declaration by value instead of reference.
This allows the sem_type of those references to keep the inferred_not_null flag, when we pop out the branch context when doing flow analysis.
Keeping this inferred flag is important when generating dummy values for cg_query_plan.