Simplify query plan generation code for shared fragments, they can have normal variables, the variables just flow.
This avoids cases where we might try to use a local variable that is not defined in the shared fragment because our variable replacement was just messing us up.  We still need to replace native calls because we can't assume the native function will be available much less initialized, but those can only happen inside the context of call to a shared proc and so they are never evaluated in a sql context.  If they ever were everything breaks, not just query plans, so we don't have to worry about that going wrong.

Added more test cases for nested fragments and also for top level fragment with a blob.  Note that if you try to simply use the trivial blob and object functions directly the top level case fails, hence the test case to ensure that isn't happening.