Added bupdatekey -- no tests yet though bupdate key is totally untested, run_test will need changes but this is the next checkpoint.
The strategy is to figure out how much extra variable space we need
for the new buffer based on the updates to variable args.  We note
all the updated fields, any duplicates cause errors.

Then we make the new resulting blob at the new size, we copy all
the fixed stuff over and then apply the updates including adding
the variable sized blobs to the end of the new blob in its empty
variable section

Finally we copy over any variable sized items (text and blobs)
that are were not updated.