Rename shared fragment names in dot expressions This is a long standing bug where if you used the name of a table parameter in the select expression you would get a run-time error in SQLite because we did not rename it.
The test case looks like this:

```
[[shared_fragment]]
proc transformer()
begin
  with
     source(*) like xy
     select source.x + 1 x, source.y + 20 y from source;
end;
```

Previously "source.x" and "source.y" would not be renamed to whatever
was passed in for the source argument (e.g. to maybe xy.x, xy.y).

This would yield an error at runtime.  Note this basically never
happened in practice because the norm looks like this:

```
[[shared_fragment]]
proc transformer()
begin
  with
     source(*) like xy
     select S.x + 1 x, S.y + 20 y from source S;
end;
```

Which was immune to the issue since S does *not* need to be renamed.