Adding foo::bar(x) support for basic overloads
It's already the case that `x:foo(...)` expands to `foo(x, ...)`
which allows you to chain calls in a pipeline.  The syntax is
borrowed from Lua

In this pull we add `::` which appends the type of the left argument.

So for instance

```
let x := true:format();
```

becomes

```
let x := format_bool(true);
```

All scalar types and cursors can be modified this way.

There is no new codegen, this is just sugar.