Allow top level expressions as statements Automatically convert top level function call expressions to procedure calls if the target is a procedure Apply the reverse argument operators before doing so therefor you can write something like `x::dump()` and get the right dumping.
NOTE: top level calls prefer procedures to functions if both exist.  This is so that
`printf("%d\n", 1);`  works even though there is a printf function.

NOTE: If the procedure has an out argument that is normally used for the result (the proc as func form)
then the top level call will claim it is missing this argument.  This is because it's been rewritten as
a call.  This can be fixed by not doing the rewrite if a trailing out argument is missing.  That isn't
done yet.

Still, at this point with the automatic rewrites you just get the right thing and you don't have
to say `call` except for the one glitch above.

Since `:=` is not an expression operator there are really no other cases where it makes sense to do
expressions at the top level.  CQL has no side effect operators like `++` or `+=`.