Handle expr_stmt calls to procs with 1 out arg This refines the previous automatic conversion of top level function calls to a procedure to call statements. The essential thing here is that if there is one missing tail argument then this is the proc_as_func case.  The out argument is treated as the result.  In that case should not rewrite the function call to a procedure call.  This will automatically create a temporary to hold the result as usual and then discard it like all statement expressions.
To do this we had to actually fully expand the arguments and count
them.  There are already helpers for all of this but it's quite
a lot of checking.  We now rewrite the arg list expanding and
"FROM shape" forms so that we can get the actual count of arguments.

Previously we didn't need it because we didn't care.  There's
now a more code for this one special case than everything else
put together on this path but it's still not very much.