Consistently use arg_list nodes for arguments This is one step towards unifying the AST for proc calls and func calls Which is different for historical reasons none of which are good. We usually don't do this stuff because it changes the test output and breaks tests in an annoying fashion but it's really stupid that we sometimes use ast_star and sometimes don't and we have to do arg stuff in two places all the time.  This change will let us change the grammar surgically in another diff.
Note I changed the error test case only because I noticed some versions
of bison emit "end of file" and others "$eof" in the error message
and that isn't what the test is about anyway.  So I picked something
that has no such weirdness and just gives a for sure syntax error.