Removing common/cov_common.sh
The common/ .sh files were there because the Meta build needed to
be different than the regular OSS build.  Supporting the Meta build
is now a non-goal so I'm pulling these back into the normal place.

Basically meta had the same cov_common.sh as OSS but a unique cov.sh.

This is just weird now.