Nullability analysis on logical AND/OR expressions
Be able to infer a variable to be non null when one does a `x IS [NOT] NULL` followed either by `AND` or an `OR`, so given:
```
let v1 := x is not null and _do_something(x);
let v2 := x is null or _do_something(x);
```
`x` would be inferred as not null when being passed to `_do_something` in both cases.

This seems fairly straightforward to support given the existing framework for flow analysis.