@include is no longer allowed except at the top level
Again reducing include loopholes.

None of this business:

proc foo()
begin
  @include "body"
end;

These restrictions of course limit what you can do with include and that's the point.
The ultimate goal is to make include be a thing that behaves like "import" not like
a bit of text you can smear everywhere.