Add builtins 'like' 'glob' and 'substring'
Substring is a trivial alias for substr.

Like and Glob are simple requiring all string args and having
normal null and sensitive semantics.  However both LIKE and GLOB
are keywords and they cannot be make into identifiers without
creating ambiguity.  So they were special cased as valid function
names.  This also clears them for use in pipeline functions with
the same grammar.  This is a bit annoying but it can't be avoided
with a bison parser.