Cleaner macros -- the tree-sitter way (#115)
* removing all the lexer and parser talk back in favor of smarter expansion
The tree sitter grammar proves that there is an unambiguous parse of the
macros without knowing the macro type in the lexer.  So we're getting rid
of that.  This also helps us to better parse disabled @ifdef bodies.  Again
the tree-sitter grammar proves we don't need this complexity.  We can figure
out what we need to do expanding by looking at the kind of macro we have
in the tree.

* Handle parsing macros on the other arm of the else this was kind of the whole point of this exercise

* The grammar now parses macros "the tree-sitter way" as a result there is no need to override all the macro stuff with something more sensible because it's already sensible.

* Additionall error cases with better messages for new macro logic