Insisting on exact binary equivalience between C and Lua blob serializer This found two issues in the C code.  * the size of blobs should have been varint encoded, it wasn't  * when a bool value as null the bool_index was not advanced, storage is present for bools even if they are null
The first of these created a system that was bigger than it needed to be but still worked.  The second made
nullable bools totally broken.  This is technically a breaking change but inasmuch as the bool issue made
it so that the old version was unusable I feel confident that nobody was using this...