The blob storage feature is now done entirely with normal functions (#121)
* Use blob storage function forms in runtime tests

* Added test coverage for the unable to infer blob type case

* Add suitable testing for error cases in C:to_blob and C:from_blob

The underlying procedures `cql_cursor_to_blob` and `cql_cursor_from_blob` have to be checked for their blob type.  There is already a helper for this.  I added a post check function in `sem.c` to make a clear hook point for such things.  If the number ever gets large we can use a table to dispatch but at present it's only two so `StrCaseCmp` is good enough.