Clean rt (#125)
* Removing most references to cqlrt->foo for simple strings
Most things that are in cqlrt->* do not need to be there
because they are used once or used inconsistently from there.
This is a bit of a hold-over from the earliest days of the
compiler when we wanted to have the source change when we
retargeed its cqlrt.c to something else.  The Meta version
had drastically different names.  It turns out the best way
to retarget cqlrt.h and so forth is to use #defines which is
what everybody did.  The compiler support to emit different
source didn't really help and it makes the code less readable
so I'm undoing most of it.  But there were good comments
for the meaning of all these things and I don't want to lose
them.  They will be a lot easier to find in cqlrt.c than
they were in cql.h buried as string values.

* Remove a bunch more rt->foo items
This is a no-op for clarity.

* remove yet more rt->foo entries that are not helpful

* We're left with a much small rt.c with things that seem to belong there