Added docs for the various @op directives that are present The idea is to record for the record what the overloads are in `cql.y` so that a person might know what they mean or even know to look there.
When I did this I discovered that :equals :hash and :diff_col and :diff_val were missing
for cursors and they are important so I added those.

While adding those I noticed that the code to compare identical function declarations
was bespoke instead of using the nromal helper which meant that no difference report
was generated if you got it wrong.

Most of the difference here is just test output due to the new builtins.