Resolving https://github.com/ricomariani/CG-SQL-author/issues/131
To resolve this we use the coding pattern that the C code uses rather than the simpler "defer the right arg by wrapping it in a function call".

The new codegen looks like this:

Note that `_tmp_n_bool_3` could be optimized out if the expression on the left already stored its result in a local
but it's harmless enough.

```
 _tmp_n_int_1 = side_effect1()
 _tmp_n_bool_3 = _tmp_int_1  -- not strictly needed
 if cql_is_true(_tmp_n_bool_3) then
   _tmp_bool_0 = true
 else
   _tmp_n_int_2 = side_effect2()
   _tmp_bool_0 = cql_logical_or(_tmp_n_bool_3, _tmp_int_2)
end
-- result is in _temp_bool_0
```

This is basically what the C code does always because it has to use `nullable_int` structs and such.

We still generate this form:

```
cql_shortcircuit_and(left_expr, function() return right_expr end))
```
If the right expression doesn't require any statements so it can't throw (i.e. there can't be a "goto" in there or anything like that.

Wrapping the right in a function lets us defer the execution to only when it is needed.  This is more economical than the most general form above.  Before this change we always used the short-circuit form, but it doesn't work if the right side might throw an exception.