Update returning (#144)
* update returning parsing and semantic validation

* added test cases for echoing update returning

* added semantic tests for update returning
found an error in backed tables, the cursor form doesn't do the backed transform

* cleaned up various sem_select forms to be the most specific
This resulted in better minification of the select list in addition to more clarity
about exactly where we are in the AST.  This is locked down with contracts.

* Added update returning to query plan output

* Added test cases for test helpers on update returning

* Added JSON tests for update returning

* returning clause refactored into one handler, added the rewrite for * and @columns in the result list

* Expand * and T.* into columns early in semantic phase for RETURNING

This gets all the column calculus working with a minimum of work.  We should do * like this in all the other cases because the current expansion strategy is inflexible. Opened a new issue for that.