Handle the `[[autodrop]]` attribute in lua Since you can now call fetch_results in CQL I ported the test case to CQL and added it as a normal run test and deleted it from run_test_client.c. I also macroized the MODERN test cases (requiring a recent SQLite) so that they are clear and can appear anywhere.  I accidentally added a non modern test case inside the @ifdef and decided that mistake is never going to be made again.  Lucky I'm paranoid and I tend to verify that my test actually ran before I assume it is passing.
Autodrop itself is trivial, the code from cg_c.c was readily adapted
to make a list of tables to drop and then a lua helper drops them
very much like in C.  This happens after the result set is fetched.
There is no good way to do this inside the proc so that's why we
have the attribute.  It's helpful if you use temp tables to
create your result. Note that autodrop can only drop temp tables
and that is (a) verified and (b) not new.