Added run tests for the remaining returning flavors. In the process of doing this we found several issues: * the upsert where clause was wrong, it doesn't need the rowid constraint * the other parts of that where clause need the column conversion for backing * we have to ensure there is a where clause for the ambiguity problem, so we add one as needed * various cases were missing `quoted identifier` handling * the various test cases were corrected in addition to adding these new ones
And that ladies and gentlemen is why we don't really believe a feature is working until we have
at least a few "run tests". I know the codegen output better than anyone and I still get it wrong.

The actual "returning" part was fine other than the fact I had the syntax wrong
all the issues were with returning combined with backed tables.