Remove --rt objc -- it is completely replaced by cqlrt_cf/cqlobjc.py (#154)
ObjC is now on the same footing as all the other ABI wrappers and indeed
the python is a lot easier for most people to work with.  I hope to
extend it to provide real classes like Java and .NET in the future.

The Meta version of CQL did this extraction a couple of years ago but
there was no open source version of the python solution so I kept it until now.