Adjusting copyright headers for accuracy
This commit begins the practice of correctly noting the origin of code in the project.
In its initial history everything was copyright by Meta;  outside contributions were
highly limited and incremental.  However since I took over as maintainer Meta contributions
have been very modest indeed and after a couple of months stopped entirely.  As a result
there were many "post Meta" files that had nothing to do with Meta.  For consistency I
maintained the same copyright banner, but it was in some sense wrong.

After consulting with other maintainers it seems that the best practice is to be
as accurate as possible on a file by file basis for copyright ownership and keep the LICENSE
uniform. I have done so.  As it happens this basically results in a bunch of files changing
to my own copyright as they had nothing to do with Meta.

I have noted this change of practice in the LICENSE file.

There may be a few more commits of this ilk as I sort out the rest of the post-meta work
for accuracy.