Include block may include @schema version In order to allow use of schema include and also declaring the upgrade version the include block is generalized to include the @upgrade_schema_version directive. In fact, this becomes the *only* place this may appear as it must appear before any meaningful code, especially table declarations.  This adds a nice new invariant to that code.
Additionally we now force ignore -Wparenthesis like many other such directives
because those warnings are for humans. CQL is not confused about order of operations.
There are many others in this category where we could make the code uglier to avoid
having the code look like a human added redundant parens and so forth. Our position
on this has been to suppress useless warnings and opt-into useful ones.  i.e. any
warning that is the kind of thing an editor of cg_c.c would get wrong.  Order of
operations is not one of these. In fact, adding special cases makes it more likely
that we will get something wrong...